+++ packages/riverpod/lib/src/state_notifier_provider/auto_dispose.dart
@@ -1,24 +1,51 @@
 part of '../state_notifier_provider.dart';
 
+class _AutoDisposeNotifierProvider<Notifier extends StateNotifier<Object?>>
+    extends AutoDisposeProvider<Notifier> {
+  _AutoDisposeNotifierProvider(
+    Create<Notifier, AutoDisposeProviderReference> create, {
+    required String? name,
+  }) : super(
+          (ref) {
+            final notifier = create(ref);
+            ref.onDispose(notifier.dispose);
+            return notifier;
+          },
+          name: name == null ? null : '$name.notifier',
+        );
+}
+
 /// {@macro riverpod.statenotifierprovider}
 @sealed
 class AutoDisposeStateNotifierProvider<Notifier extends StateNotifier<Value>,
-    Value> extends AutoDisposeProviderBase<Notifier, Value> {
+        Value> extends AutoDisposeProviderBase<Notifier, Value>
+    with _StateNotifierProviderMixin<Notifier, Value> {
   /// {@macro riverpod.statenotifierprovider}
-  AutoDisposeStateNotifierProvider(
-    Create<Notifier, AutoDisposeProviderReference> create, {
-    String? name,
-  }) : super(create, name);
-
-  // TODO name
-  late final RootProvider<Notifier, Notifier> notifier =
-      AutoDisposeCreatedProvider(this);
+  AutoDisposeStateNotifierProvider(this._create, {String? name}) : super(name);
 
   /// {@macro riverpod.family}
   static const family = AutoDisposeStateNotifierProviderFamilyBuilder();
 
   /// {@macro riverpod.autoDispose}
   static const autoDispose = AutoDisposeStateNotifierProviderBuilder();
+
+  final Create<Notifier, AutoDisposeProviderReference> _create;
+
+  @override
+  AutoDisposeStateNotifierProviderFamily<Notifier, Value, Object?>? get from =>
+      super.from
+          as AutoDisposeStateNotifierProviderFamily<Notifier, Value, Object?>?;
+
+  // TODO name
+  @override
+  late final AutoDisposeProviderBase<Notifier, Notifier> notifier = from != null
+      ? from!._notifierFamily(argument)
+      : _AutoDisposeNotifierProvider(_create, name: name);
+
+  @override
+  Notifier create(covariant AutoDisposeProviderReference ref) {
+    return ref.watch(notifier);
+  }
 
   @override
   _StateNotifierProviderState<Notifier, Value> createState() {
@@ -47,6 +74,45 @@ class AutoDisposeStateNotifierProviderFamily<
     String? name,
   ) {
     return AutoDisposeStateNotifierProvider(
+      (ref) => builder(ref, value),
+      name: name,
+    );
+  }
+
+  late final _notifierFamily =
+      _AutoDisposeNotifierFamily<Notifier, Param>(builder, name);
+
+  /// Overrides the behavior of a family for a part of the application.
+  ///
+  /// {@macro riverpod.overideWith}
+  Override overrideWithProvider(
+    Notifier Function(ProviderReference ref, Param param) builderOverride,
+  ) {
+    return FamilyOverride(
+      _notifierFamily,
+      (param) {
+        return _notifierFamily.create(param as Param, builderOverride, null);
+      },
+    );
+  }
+}
+
+@sealed
+class _AutoDisposeNotifierFamily<Notifier extends StateNotifier<Object?>, Param>
+    extends Family<Notifier, Notifier, Param, AutoDisposeProviderReference,
+        _AutoDisposeNotifierProvider<Notifier>> {
+  _AutoDisposeNotifierFamily(
+    Notifier Function(AutoDisposeProviderReference ref, Param param) builder,
+    String? name,
+  ) : super(builder, name);
+
+  @override
+  _AutoDisposeNotifierProvider<Notifier> create(
+    Param value,
+    Notifier Function(AutoDisposeProviderReference ref, Param param) builder,
+    String? name,
+  ) {
+    return _AutoDisposeNotifierProvider(
       (ref) => builder(ref, value),
       name: name,
     );