+++ packages/riverpod/test/provider/state_notifier_provider_test.dart
@@ -6,74 +6,69 @@ import '../utils.dart';
 
 void main() {
   test('StateNotifierFamily override', () {
-    final provider = StateNotifierProvider.family<TestNotifier, int>(
-        (ref, a) => TestNotifier());
-    final notifier2 = TestNotifier(42);
-    final container = ProviderContainer(
-      overrides: [provider.overrideWithProvider((ref, a) => notifier2)],
-    );
-
-    // access in the child container
-    // try to read provider.state before provider and see if it points to the override
-    expect(container.read(provider(0).state), 42);
-    expect(container.read(provider(0)), notifier2);
-  });
-
-  test('can be assigned to provider', () {
-    final Provider<TestNotifier> provider = StateNotifierProvider((_) {
+    final provider =
+        StateNotifierProvider.family<TestNotifier, int, int>((ref, a) {
       return TestNotifier();
     });
-    final container = ProviderContainer();
+    final notifier2 = TestNotifier(42);
+    final container = ProviderContainer(
+      overrides: [
+        provider.overrideWithProvider((ref, a) => notifier2),
+      ],
+    );
 
-    expect(container.read(provider), isA<TestNotifier>());
+    expect(container.read(provider(0).notifier), notifier2);
+    // access in the child container
+    // try to read provider.state before provider and see if it points to the override
+    expect(container.read(provider(0)), 42);
   });
 
   test('overriding the provider overrides provider.state too', () {
     final notifier = TestNotifier(42);
-    final provider = StateNotifierProvider((_) => TestNotifier());
+    final provider =
+        StateNotifierProvider<TestNotifier, int>((_) => TestNotifier());
     final container = ProviderContainer(
       overrides: [
-        provider.overrideWithProvider(
-            StateNotifierProvider((_) => TestNotifier(10)))
+        provider.overrideWithValue(TestNotifier(10)),
       ],
     );
 
     // does not crash
     container.updateOverrides([
-      provider.overrideWithProvider(StateNotifierProvider((_) => notifier)),
+      provider.overrideWithValue(notifier),
     ]);
 
-    expect(container.read(provider), notifier);
-    expect(container.read(provider.state), 42);
+    expect(container.read(provider.notifier), notifier);
+    expect(container.read(provider), 42);
 
     notifier.increment();
 
-    expect(container.read(provider.state), 43);
+    expect(container.read(provider), 43);
   });
 
   test('can specify name', () {
-    final provider = StateNotifierProvider(
-      (_) => TestNotifier(),
-      name: 'example',
-    );
+    final provider = StateNotifierProvider<TestNotifier, int>((_) {
+      return TestNotifier();
+    }, name: 'example');
 
+    expect(provider.notifier.name, 'example.notifier');
     expect(provider.name, 'example');
-    expect(provider.state.name, 'example.state');
 
-    final provider2 = StateNotifierProvider((_) => TestNotifier());
+    final provider2 =
+        StateNotifierProvider<TestNotifier, int>((_) => TestNotifier());
 
+    expect(provider2.notifier.name, isNull);
     expect(provider2.name, isNull);
-    expect(provider2.state.name, isNull);
   });
 
   test('disposes the notifier when provider is unmounted', () {
     final notifier = TestNotifier();
-    final provider = StateNotifierProvider<TestNotifier>((_) {
+    final provider = StateNotifierProvider<TestNotifier, int>((_) {
       return notifier;
     });
     final container = ProviderContainer();
 
-    expect(container.read(provider), notifier);
+    expect(container.read(provider.notifier), notifier);
     expect(notifier.mounted, isTrue);
 
     container.dispose();
@@ -83,13 +78,13 @@ void main() {
 
   test('provider subscribe the callback is never', () async {
     final notifier = TestNotifier();
-    final provider = StateNotifierProvider<TestNotifier>((_) {
+    final provider = StateNotifierProvider<TestNotifier, int>((_) {
       return notifier;
     });
     final listener = ControllerListenerMock();
     final container = ProviderContainer();
 
-    final sub = provider.addLazyListener(
+    final sub = provider.notifier.addLazyListener(
       container,
       mayHaveChanged: () {},
       onChange: listener,
@@ -112,13 +107,13 @@ void main() {
   });
 
   test('provider subscribe callback never called', () async {
-    final provider = StateNotifierProvider<TestNotifier>((_) {
+    final provider = StateNotifierProvider<TestNotifier, int>((_) {
       return TestNotifier();
     });
     final listener = ListenerMock();
     final container = ProviderContainer();
 
-    final sub = provider.state.addLazyListener(
+    final sub = provider.addLazyListener(
       container,
       mayHaveChanged: () {},
       onChange: listener,
@@ -127,7 +122,7 @@ void main() {
     verify(listener(0)).called(1);
     verifyNoMoreInteractions(listener);
 
-    container.read(provider).increment();
+    container.read(provider.notifier).increment();
 
     verifyNoMoreInteractions(listener);
     sub.read();
@@ -145,6 +140,11 @@ class TestNotifier extends StateNotifier<int> {
   TestNotifier([int initialValue = 0]) : super(initialValue);
 
   void increment() => state++;
+
+  @override
+  String toString() {
+    return 'TestNotifier($state)';
+  }
 }
 
 class ListenerMock extends Mock {